+++ include/reactphysics3d/configuration.h
@@ -129,6 +129,9 @@ constexpr uint16 NB_MAX_CONTACT_POINTS_IN_POTENTIAL_MANIFOLD = 256;
 /// Distance threshold to consider that two contact points in a manifold are the same
 constexpr decimal SAME_CONTACT_POINT_DISTANCE_THRESHOLD = decimal(0.01);
 
+/// Global alignment (in bytes) that all allocators must enforce
+constexpr uint8 GLOBAL_ALIGNMENT = 16;
+
 /// Current version of ReactPhysics3D
 const std::string RP3D_VERSION = std::string("0.9.0");