+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Changed
 
+ - The library must now be compiled with C++ 17 compiler
+ - If the user sets its custom allocator, the return allocated memory must now be 16 bytes aligned
+
 ### Removed
 
 ### Fixed